Рефактор по interaction-api

+++ core/user-service/src/main/java/ru/practicum/user/param/UserParams.java
@@ -8,7 +8,7 @@ import java.util.List;
 @Data
 @AllArgsConstructor
 public class UserParams {
-    private List<Integer> ids;
+    private List<Long> ids;
     private Integer from;
     private Integer size;
 }
